update aop log

+++ Server-Platform/Hibernate/3.0/Code/BPlatform3.0/src/main/java/com/memory/platform/modules/system/base/service/ISystemLogService.java
@@ -10,5 +10,5 @@ public interface ISystemLogService extends IBaseService<SystemLog>{
 	 * @param loglevel 级别
 	 * @param operatetype 类型
 	 */
-	public void addLog(String LogContent, Short loglevel,Short operatetype);
+	public void addLog(String LogContent, int loglevel,int operatetype);
 }
